src/vlcsync-client/requirements.txt
python-vlc==3.*
PyQt5==5.15.9